Divided the controllers by responsibility and added a controller for the admin role.
