+++ Restaraunt.WebApi/Controllers/AccountController.cs
@@ -1,12 +1,8 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using Restaraunt.Application.Extensions;
-using Restaraunt.Application.Interfaces;
-using Restaraunt.Domain.Entities;
 using Restaraunt.Domain.Entities.Identity;
-using Restaraunt.Persistence;
 using Restaraunt.WebApi.Models.Identity;
 using System.IdentityModel.Tokens.Jwt;
 
@@ -14,101 +10,15 @@ namespace Restaraunt.WebApi.Controllers
 {
 	[AllowAnonymous]
 	[Route("api/[controller]")]
-	[ApiController]
 	public class AccountController : BaseController
 	{
 		private readonly UserManager<User> _userManager;
-		private readonly ProductDbContext _context;
-		private readonly ITokenService _tokenService;
 		private readonly IConfiguration _configuration;
 
-		public AccountController(UserManager<User> userManager, ProductDbContext context,
-			ITokenService tokenService, IConfiguration configuration)
+		public AccountController(UserManager<User> userManager, IConfiguration configuration)
 		{
 			_userManager = userManager;
-			_context = context;
-			_tokenService = tokenService;
 			_configuration = configuration;
-		}
-
-		[HttpPost("login")]
-		public async Task<ActionResult<AuthResponse>> Login([FromBody] AuthRequest request)
-		{
-			if (!ModelState.IsValid)
-			{
-				return BadRequest(ModelState);
-			}
-
-			var managedUser = await _userManager.FindByEmailAsync(request.Email);
-
-			if (managedUser == null)
-			{
-				return BadRequest("Bad credentials");
-			}
-
-			var isPasswordValid = await _userManager.CheckPasswordAsync(managedUser, request.Password);
-
-			if (!isPasswordValid)
-			{
-				return BadRequest("Bad credentials");
-			}
-
-			var user = _context.Users.FirstOrDefault(u => u.Email == request.Email);
-
-			if (user is null)
-				return Unauthorized();
-
-			var roleIds = await _context.UserRoles.Where(r => r.UserId == user.Id).Select(x => x.RoleId).ToListAsync();
-			var roles = _context.Roles.Where(x => roleIds.Contains(x.Id)).ToList();
-
-			var accessToken = _tokenService.CreateToken(user, roles);
-			user.RefreshToken = _configuration.GenerateRefreshToken();
-			user.RefreshTokenExpiryTime = DateTime.UtcNow.AddDays(_configuration.GetSection("Jwt:RefreshTokenValidityInDays").Get<int>());
-
-			await _context.SaveChangesAsync();
-
-			return Ok(new AuthResponse
-			{
-				Username = user.UserName!,
-				Email = user.Email!,
-				Token = accessToken,
-				RefreshToken = user.RefreshToken
-			});
-		}
-
-		[HttpPost("register")]
-		public async Task<ActionResult<AuthResponse>> Register([FromBody] RegisterRequest request)
-		{
-			if (!ModelState.IsValid) return BadRequest(request);
-
-			var user = new User
-			{
-				Email = request.Email,
-				UserName = request.UserName
-			};
-			var result = await _userManager.CreateAsync(user, request.Password);
-
-			foreach (var error in result.Errors)
-			{
-				ModelState.AddModelError(string.Empty, error.Description);
-			}
-
-			if (!result.Succeeded) return BadRequest(request);
-
-			var findUser = await _context.Users.FirstOrDefaultAsync(x => x.Email == request.Email);
-
-			if (findUser == null) throw new Exception($"User {request.Email} not found");
-
-			var cart = new Cart() { UserId = findUser.Id };
-			await _context.Carts.AddAsync(cart);
-
-			await _userManager.AddToRoleAsync(findUser, RoleConsts.Customer);
-
-			return await Login(new AuthRequest
-			{
-				Email = request.Email,
-				Password = request.Password
-			});
 		}
 
 		[HttpPost]
@@ -150,7 +60,7 @@ namespace Restaraunt.WebApi.Controllers
 			});
 		}
 
-		[Authorize]
+		[Authorize(Roles = "Administrator")]
 		[HttpPost]
 		[Route("revoke/{username}")]
 		public async Task<IActionResult> Revoke(string username)
