+++ Restaraunt.WebApi/Models/Identity/RoleConsts.cs
@@ -3,6 +3,6 @@
 	public class RoleConsts
 	{
 		public const string Customer = "Customer";
-		public const string Administrator = "Administrator";
+		public const string Administrator = "Admin";
 	}
 }
